Add wrapper component
